+++ demo/actions.ts
@@ -1,12 +1,11 @@
-import { a, p } from '../src/a'
+import a from '../lib/a'
 
-const ATestAction = a('A_TEST_ACTION', p<{ str: string }>())
-type ATestAction = ReturnType<typeof ATestAction>
-
+const ATestAction = a('A_TEST_ACTION', <{ str: string }>{})
 const ASecondTestAction = a('A_SECOND_TEST_ACTION')
-type ASecondTestAction = ReturnType<typeof ASecondTestAction>
+const AThirdTestAction = a('A_THIRD_TEST_ACTION', <{ num: number }>{})
 
-const AThirdTestAction = a('A_THIRD_TEST_ACTION', p<{ num: number }>())
+type ATestAction = ReturnType<typeof ATestAction>
+type ASecondTestAction = ReturnType<typeof ASecondTestAction>
 type AThirdTestAction = ReturnType<typeof AThirdTestAction>
 
 export { ATestAction, ASecondTestAction, AThirdTestAction }